created two api calls for latest price
